# control-plane/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Control-plane is a central microservice of Service Mesh which is responsible for all the gateways configuration. For more information, visit our Documentation(https://github.com/Netcracker/qubership-core-control-plane/tree/main/README.md).",
        "title": "Control Plane API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v3/active-active": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Active Config Post",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Active Config Post",
                "operationId": "ActiveActiveConfigPost",
                "parameters": [
                    {
                        "description": "ActiveDCsV3",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ActiveDCsV3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Active Config Delete",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Active Config Delete",
                "operationId": "ActiveActiveConfigDelete",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v3/apply-config": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Post Apply Config",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Post Apply Config",
                "operationId": "HandlePostConfig",
                "parameters": [
                    {
                        "description": "ConfigResource",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configresources.ConfigResource"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v3.ApplyResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorcodes.CpErrCodeError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorcodes.CpErrCodeError"
                        }
                    }
                }
            }
        },
...
        "/api/v3/config": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Apply Config",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Apply Config",
                "operationId": "HandleConfig",
                "parameters": [
                    {
                        "description": "ConfigResource",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configresources.ConfigResource"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v3.ApplyResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorcodes.CpErrCodeError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorcodes.CpErrCodeError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "configresources.ConfigResource": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/configresources.Metadata"
                },
                "nodeGroup": {
                    "type": "string"
                },
                "spec": {}
            }
        },
        "configresources.Metadata": {
            "type": "object",
            "additionalProperties": true
        },
...
        "dto.ActiveDCsCommonLbConfigV3": {
            "type": "object",
            "properties": {
                "healthyPanicThreshold": {
                    "type": "number"
                }
            }
        },
        "dto.ActiveDCsHealthCheckV3": {
            "type": "object",
            "properties": {
                "healthyThreshold": {
                    "type": "integer"
                },
                "interval": {
                    "type": "integer"
                },
                "noTrafficInterval": {
                    "type": "integer"
                },
                "timeout": {
                    "type": "integer"
                },
                "unhealthyInterval": {
                    "type": "integer"
                },
                "unhealthyThreshold": {
                    "type": "integer"
                }
            }
        },
        "dto.ActiveDCsRetryBackOffV3": {
            "type": "object",
            "properties": {
                "baseInterval": {
                    "type": "integer"
                },
                "maxInterval": {
                    "type": "integer"
                }
            }
        },
        "dto.ActiveDCsRetryPolicyV3": {
            "type": "object",
            "properties": {
                "numRetries": {
                    "type": "integer"
                },
                "perTryTimeout": {
                    "type": "integer"
                },
                "retriableStatusCodes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "retryBackOff": {
                    "$ref": "#/definitions/dto.ActiveDCsRetryBackOffV3"
                },
                "retryOn": {
                    "type": "string"
                }
            }
        },
        "dto.ActiveDCsV3": {
            "type": "object",
            "properties": {
                "commonLbConfig": {
                    "$ref": "#/definitions/dto.ActiveDCsCommonLbConfigV3"
                },
                "healthCheck": {
                    "$ref": "#/definitions/dto.ActiveDCsHealthCheckV3"
                },
                "httpPort": {
                    "type": "integer"
                },
                "httpsPort": {
                    "type": "integer"
                },
                "privateGwHosts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "protocol": {
                    "type": "string"
                },
                "publicGwHosts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "retryPolicy": {
                    "$ref": "#/definitions/dto.ActiveDCsRetryPolicyV3"
                }
            }
        },
...
        "errorcodes.CpErrCodeError": {
            "type": "object",
            "properties": {
                "cause": {},
                "detail": {
                    "type": "string"
                },
                "errorCode": {
                    "$ref": "#/definitions/errors.ErrorCode"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stackTrace": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "errors.ErrorCode": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "v3.ApplyResult": {
            "type": "object",
            "properties": {
                "request": {
                    "$ref": "#/definitions/configresources.ConfigResource"
                },
                "response": {
                    "$ref": "#/definitions/v3.HandlingResponse"
                }
            }
        },
        "v3.HandlingResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}